etl/requirements.txt
pydantic==1.9.0
psycopg2-binary==2.9.3
elasticsearch===7.14.0
